MPV - notebooks and logging work for basic MAP and signal parsing

+++ ep_parse/case_data.py
@@ -1,6 +1,5 @@
 import os
 import json
-import toolz as tz
 import yaml
 import pandas as pd
 import logging
@@ -10,11 +9,10 @@ from dotenv import load_dotenv
 
 load_dotenv()
 
-import ep_parse.utils as pu
-import ep_parse.constants as pc
+import ep_parse.utils as u
 import ep_parse.yaml_helper as ymlh
 import ep_parse.common as pic
-import ep_parse.tag as tg
+from ep_parse.common import DataSource
 
 log = logging.getLogger(__name__)
 
@@ -24,14 +22,25 @@ class FileType(StrEnum):
     TAGS = "tags.json"
     EPSYSTEM_EVENTS = "epsystem_events.csv"
     SIGNALS = "signals.h5"
-    EPLOG = "epmed_bookmarks.txt"
+    EPMED_LOG = "epmed_bookmarks.txt"
     CARDIOLAB_LOG = "cardiolab_logs.txt"
 
 
-def case_file_path(case_id: str, file_type: FileType | pic.DataSource) -> str:
-    return os.path.join(
-        os.environ["data_filepath"], case_id, (f"{case_id}_" if isinstance(file_type, FileType) else "") + file_type
-    )
+def case_file_path(case_id: str, file_type: FileType | DataSource = None) -> str:
+    """Returns a filepath associated with a case, based on filetype or Datasource
+
+    Args:
+        case_id (str): The case of interest
+        file_type (FileType | DataSource, optional): file (FileType), subdirectory (datasource), or case directory (None). Defaults to None.
+
+    Returns:
+        str: path to the file or directory of interest
+    """
+    cpath = ""
+    if file_type:
+        cpath = (f"{case_id}_" if isinstance(file_type, FileType) else "") + file_type
+
+    return os.path.join(os.environ["data_filepath"], case_id, cpath)
 
 
 def case_signals_db(case_id: str, mode: str = "r", compression: bool = True) -> pd.HDFStore:
@@ -55,7 +64,19 @@ def case_PPS(case_id: str):
     with case_signals_db(case_id) as s:
         for g in s.keys():
             if ("signals/") in g:
-                return pu.points_per_second(s.select(g, stop=2).index)
+                return u.points_per_second(s.select(g, stop=2).index)
+
+
+def get_MAPPING_system(case_id: str) -> pic.MAPPING_SYSTEM:
+    _meta = load_case_meta(case_id)
+    _meta = _meta if "mapping_system" in _meta else infered_case_properties(case_id)
+    return _meta.get("mapping_system")
+
+
+def get_EP_system(case_id: str) -> pic.EP_SYSTEM:
+    _meta = load_case_meta(case_id)
+    _meta = _meta if "ep_system" in _meta else infered_case_properties(case_id)
+    return _meta.get("ep_system")
 
 
 def case_channels(case_id: str):
@@ -73,16 +94,9 @@ def store_pps_in_meta(case_id: str) -> int:
     update_case_meta(case_id, ["signal_sample_rate_hz"], pps)
 
 
-def create_case_meta(case_id: str) -> None:
-    meta_file = case_file_path(case_id, FileType.META)
-    if os.path.exists(meta_file):
-        return
+def infered_case_properties(case_id: str) -> dict:
     dir_path = os.path.join(os.environ["data_filepath"], case_id)
-    if not os.path.exists(dir_path):
-        os.mkdir(dir_path)
 
-    # with open("resources/default_meta_attributes.json", "r") as fp:
-    #     meta_data = json.load(fp)
     meta_data = {"case_id": case_id}
 
     for epsys in pic.EP_SYSTEM:
@@ -97,6 +111,19 @@ def create_case_meta(case_id: str) -> None:
 
     meta_data["md_catheter?"] = os.path.exists(os.path.join(dir_path, "md_catheter_logs"))
 
+    return meta_data
+
+
+def create_case_meta(case_id: str) -> None:
+    meta_file = case_file_path(case_id, FileType.META)
+    if os.path.exists(meta_file):
+        return
+    meta_data = infered_case_properties(case_id)
+
+    dir_path = os.path.join(os.environ["data_filepath"], case_id)
+
+    if not os.path.exists(dir_path):
+        os.mkdir(dir_path)
     with open(meta_file, "w") as fp:
         yaml.dump(dict(sorted(meta_data.items())), fp, indent=2)
 
@@ -112,19 +139,13 @@ def load_case_tags(case_id: str):
 
 
 def write_case_tags(case_id: str, tags: list[dict], mode: str = "w"):
-    "modes: a = append, w = write (truncate), r = replace (by type)"
+    "modes: a = append, w = write (truncate)"
     if case_id:
         tags_file = case_file_path(case_id, FileType.TAGS)
-        if not os.path.exists(tags_file):
-            mode = "w"
 
-        if os.path.exists(tags_file):
+        if os.path.exists(tags_file) and mode == "a":
             old_tags = load_case_tags(case_id)
-            if mode == "a":
-                tags = old_tags + tags
-            elif mode == "r":
-                replace_type = tg.tag_type(tags[0])
-                tags = tags + [t for t in old_tags if tg.tag_type(t) != replace_type]
+            tags = old_tags + tags
 
         with open(tags_file, "w") as fp:
             json.dump(list(sorted(tags, key=itemgetter("start_time"))), fp, indent=2)
@@ -154,11 +175,6 @@ def load_case_meta(case_id: str) -> dict:
     else:
         meta = {}
 
-    meta["feature_detection_params"] = pu.merge_with_overrides(
-        pc.DEFAULT_FEATURE_DETECTION_PARAMS,
-        meta.get("feature_detection_params"),
-    )
-    meta = tz.assoc_in(meta, ["feature_detection_params", "qrs"], meta.get("qrs_meta"))
     return meta
 
 
@@ -182,7 +198,7 @@ def load_local_events(case_id: str, rfs_only: bool = False) -> pd.DataFrame:
         event_df = pd.DataFrame(tags)
         event_df = event_df[event_df.columns.intersection(["label", "start_time", "end_time", "notes"])]
         for f in ["start_time", "end_time"]:
-            event_df[f] = event_df[f].map(lambda s: pu.as_datetime(pu.as_time_str(s)) if s else pd.NaT)
+            event_df[f] = event_df[f].map(lambda s: u.as_datetime(u.as_time_str(s)) if s else pd.NaT)
         event_df.rename(columns={"label": "event"}, inplace=True)
         if rfs_only:
             event_df = event_df[event_df["event"].str.startswith("RF")]
